Added cleanup for parent tmp files in HybridNew
Also removed annoying complaint that tmp file could not be deleted when
running with HybridNew even though it was?!

FIXME - Should look into why return status from unlink was -1 in this case
and revert back to warning if error occurs
