+++ interface/Combine.h
@@ -1,7 +1,9 @@
 #ifndef HiggsAnalysis_CombinedLimit_Combine_h
 #define HiggsAnalysis_CombinedLimit_Combine_h
 #include <TString.h>
+#include <TFile.h>
 #include <boost/program_options.hpp>
+#include <boost/filesystem.hpp>
 #include "RooArgSet.h"
 #include "RooAbsReal.h"
 #include <boost/algorithm/string/split.hpp>
@@ -30,6 +32,20 @@ extern  std::string setPhysicsModelParameterExpression_;
 extern  std::string setPhysicsModelParameterRangeExpression_;
 extern  std::string defineBackgroundOnlyModelParameterExpression_;
 
+namespace { 
+    struct ToCleanUp {
+        TFile *tfile; std::string file, path;
+        ToCleanUp() : tfile(0), file(""), path("") {}
+        ~ToCleanUp() {
+              if (tfile) { tfile->Close(); delete tfile; }
+              if (!file.empty()) {  
+                 unlink(file.c_str());  // FIXME, we should check that the file deleted safely but currently when running HybridNew, we get a status of -1 even though the file is in fact removed?!
+		 //if (unlink(file.c_str()) == -1) std::cerr << "Failed to delete temporary file " << file << ": " << strerror(errno) << std::endl;
+              }
+              if (!path.empty()) {  boost::filesystem::remove_all(path); }
+        }
+    };
+}
 
 class Combine {
 public:
